pull-request: catalog the PR body rules
Split validate-body.ts into resolve-body.ts (shell parsing: heredocs,
flag values, body files, cd tracking, the title) and body-rules.ts (a
catalog of rules, each with an id, tier, detector, and message, run by
one runner that groups by tier and bundles the decision). Prose
measurements the eval scorer shares move to prose.ts, and the git probes
for repo ownership and commit existence move to repo.ts. The hook entry
now owns the orchestration in processInput and guards main() on
import.meta.main so tests can import it.

The unreadable-body deny, the backticked-commit verification, and the
personal-repo length gate become catalog entries that read their facts
from the context, which carries lazy probes so the git lookups still run
only when a body warrants them. The eval scorer defines its own
SentenceHeading shape rather than importing a rule-private type.

Tests split by module. The rule tests are one table per tier of body
fixtures and the rule ids expected to fire, with the existing fixtures
kept as rows. The runner is covered by snapshots of the bundled output.
